requirements.txt
finvizfinance>=0.14.0
yfinance>=0.2.36
fredapi>=0.5.0
jinja2>=3.1.2
python-dotenv>=1.0.0
schedule>=1.2.1
requests>=2.31.0
pandas>=2.1.0
beautifulsoup4>=4.12.2
lxml>=5.1.0
